requirements.txt
Flask
Flask-SocketIO
eventlet
pyserial
opencv-python-headless
